requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
discord==1.0.1
discord.py==1.3.4
docutils==0.15.2
idna==2.10
idna-ssl==1.1.0
jmespath==0.10.0
multidict==4.7.6
praw==7.1.0
prawcore==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-settings==0.2.2
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.2
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
yarl==1.4.2
